Update docs and use correct type for loop indices

+++ deps/ledger-zxlib/src/bignum.c
@@ -16,7 +16,6 @@
 #include "zxtypes.h"
 #include "bignum.h"
 
-
 bool_t bignumLittleEndian_bcdprint(char *outBuffer, uint16_t outBufferLen,
                                    const uint8_t *inBCD, uint16_t inBCDLen) {
     static const char hexchars[] = "0123456789ABCDEF";
@@ -32,7 +31,7 @@ bool_t bignumLittleEndian_bcdprint(char *outBuffer, uint16_t outBufferLen,
         return bool_false;
     }
 
-    for (uint8_t i = 0; i < inBCDLen; i++, inBCD++) {
+    for (uint16_t i = 0; i < inBCDLen; i++, inBCD++) {
         if (started || *inBCD != 0) {
             if (started || (*inBCD >> 4u) != 0) {
                 *outBuffer = hexchars[*inBCD >> 4u];
@@ -56,7 +55,7 @@ void bignumLittleEndian_to_bcd(uint8_t *bcdOut, uint16_t bcdOutLen,
     MEMZERO(bcdOut, bcdOutLen);
 
     uint8_t carry;
-    for (uint16_t bitIdx = 0; bitIdx < binValueLen * 8; bitIdx++) {
+    for (uint32_t bitIdx = 0; bitIdx < binValueLen * 8; bitIdx++) {
         // Fix bcd
         for (uint16_t j = 0; j < bcdOutLen; j++) {
             if ((bcdOut[j] & 0x0Fu) > 0x04u) {
@@ -122,7 +121,7 @@ void bignumBigEndian_to_bcd(uint8_t *bcdOut, uint16_t bcdOutLen,
     MEMZERO(bcdOut, bcdOutLen);
 
     uint8_t carry;
-    for (uint16_t bitIdx = 0; bitIdx < binValueLen * 8; bitIdx++) {
+    for (uint32_t bitIdx = 0; bitIdx < binValueLen * 8; bitIdx++) {
         // Fix bcd
         for (uint16_t j = 0; j < bcdOutLen; j++) {
             if ((bcdOut[j] & 0x0Fu) > 0x04u) {
